requirements.txt
appdirs==1.4.3
argcomplete==1.9.3
attrs==19.3.0
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
coolsms-python-sdk==2.0.3
Django==2.1.11
django-cors-headers==3.2.0
djangorestframework==3.10.3
docutils==0.15.2
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MLB-StatsAPI==0.1.5
pathspec==0.6.0
placebo==0.9.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.2
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.7
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
zappa==0.48.2
